adicionando permissões ao usuário pelo token

+++ MP.ApiDotNet6.Api/Controllers/PersonController.cs
@@ -2,6 +2,7 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using MP.ApiDotNet6.Application.DTOs;
 using MP.ApiDotNet6.Application.Services.Interfaces;
+using MP.ApiDotNet6.Domain.Authentication;
 using MP.ApiDotNet6.Domain.FiltersDb;
 
 namespace MP.ApiDotNet6.Api.Controllers;
@@ -10,18 +11,27 @@ namespace MP.ApiDotNet6.Api.Controllers;
 [Route("api/[controller]")]
 [ApiController]
 
-public class PersonController : ControllerBase
+public class PersonController : BaseController
 {
     private readonly IPersonService _personService;
+    private readonly ICurrentUser _currentUser;
+    private List<string> _permissionNeeded = new List<string>(){"Admin"};
+    private readonly List<string> _permissionUser;
 
-    public PersonController(IPersonService personService)
+    public PersonController(IPersonService personService, ICurrentUser currentUser)
     {
         _personService = personService;
+        _currentUser = currentUser;
+        _permissionUser = _currentUser?.Permissions?.Split(",")?.ToList() ?? new List<string>();
     }
     
     [HttpPost]
-    public async Task<ActionResult> PostAsync([FromBody] PersonDTO personDto)
+    public async Task<IActionResult> PostAsync([FromBody] PersonDTO personDto)
     {
+        _permissionNeeded.Add("CadastroPessoa");
+        if (!ValidPermission(_permissionUser, _permissionNeeded))
+            return Forbidden();
+        
         var result = await _personService.CreateAsync(personDto);
         if (result.IsSucsess)
             return Ok(result);
@@ -30,8 +40,12 @@ public class PersonController : ControllerBase
     }
 
     [HttpGet]
-    public async Task<ActionResult> GetAsync()
+    public async Task<IActionResult> GetAsync()
     {
+        _permissionNeeded.Add("BuscaPessoa");
+        if (!ValidPermission(_permissionUser, _permissionNeeded))
+            return Forbidden();
+        
         var result = await _personService.GetAsync();
         if (result.IsSucsess)
             return Ok(result);
@@ -40,8 +54,12 @@ public class PersonController : ControllerBase
     }
 
     [HttpGet("{id:int}")]
-    public async Task<ActionResult> GetByIdAsync(int id)
+    public async Task<IActionResult> GetByIdAsync(int id)
     {
+        _permissionNeeded.Add("BuscaPessoa");
+        if (!ValidPermission(_permissionUser, _permissionNeeded))
+            return Forbidden();
+        
         var result = await _personService.GetByIdAsync(id);
         if (result.IsSucsess)
             return Ok(result);
@@ -50,8 +68,12 @@ public class PersonController : ControllerBase
     }
     
     [HttpPut]
-    public async Task<ActionResult> UpdateAsync([FromBody] PersonDTO personDto)
+    public async Task<IActionResult> UpdateAsync([FromBody] PersonDTO personDto)
     {
+        _permissionNeeded.Add("EditaPessoa");
+        if (!ValidPermission(_permissionUser, _permissionNeeded))
+            return Forbidden();
+        
         var result = await _personService.UpdateAsync(personDto);
         if (result.IsSucsess)
             return Ok(result);
@@ -61,8 +83,12 @@ public class PersonController : ControllerBase
     
     [HttpDelete]
     [Route("{id}")]
-    public async Task<ActionResult> DeleteAsync(int id)
+    public async Task<IActionResult> DeleteAsync(int id)
     {
+        _permissionNeeded.Add("DeletaPessoa");
+        if (!ValidPermission(_permissionUser, _permissionNeeded))
+            return Forbidden();
+        
         var result = await _personService.DeleteAsync(id);
         if (result.IsSucsess)
             return Ok(result);
